requirements.txt
flask==3.0.0
dlib>=20.0.0
numpy>=1.26.0,<2.0.0
opencv-python>=4.8.0
pandas>=2.0.0
scikit-image>=0.22.0
scikit-learn>=1.3.0
scipy>=1.11.0
imutils==0.5.4
PyQt6>=6.6.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
